+++ apps/frontend/src/components/ErrorFallback.tsx
@@ -3,7 +3,7 @@ import './ErrorFallback.css'
 import { Button, Grid, Typography } from '@mui/material'
 import { Component, ComponentType, useEffect, useState } from 'react'
 
-import controller from '../services/controller'
+import controller from '~/services/controller'
 
 const RELOAD_COUNTDOWN = 10 // 10 second countdown before automatic reload
