+++ apps/frontend/src/components/GlobalHotKeys.tsx
@@ -2,7 +2,7 @@ import { mapValues } from 'lodash'
 import { configure, GlobalHotKeys as HotKeys } from 'react-hotkeys/es'
 import KeyEventManager from 'react-hotkeys/es/lib/KeyEventManager'
 
-import { KeyMap, mapPlatformKeys } from '../helpers/utils'
+import { KeyMap, mapPlatformKeys } from '~/helpers/utils'
 
 /**
  * React hotkeys activates `g g` twice when pressing `g g g`,
