+++ apps/frontend/src/features/controller/Navigator.tsx
@@ -19,16 +19,17 @@ import { invert } from 'lodash'
 import { useCallback, useContext, useEffect, useMemo, useState } from 'react'
 import { Navigate, useLocation } from 'react-router-dom'
 
-import { getJumpLines, getNextJumpLine } from '../lib/auto-jump'
-import { SEARCH_URL } from '../lib/consts'
-import { ContentContext, HistoryContext } from '../lib/contexts'
-import controller from '../lib/controller'
-import { useCurrentLines } from '../lib/hooks'
-import { LINE_HOTKEYS } from '../lib/keyMap'
-import { findLineIndex } from '../lib/line'
-import GlobalHotKeys from '../shared/GlobalHotKeys'
-import { withNavigationHotkeys } from '../shared/NavigationHotkeys'
-import NavigatorHotKeys from '../shared/NavigatorHotkeys'
+import GlobalHotKeys from '~/components/GlobalHotKeys'
+import { withNavigationHotkeys } from '~/components/NavigationHotkeys'
+import NavigatorHotKeys from '~/components/NavigatorHotkeys'
+import { getJumpLines, getNextJumpLine } from '~/helpers/auto-jump'
+import { SEARCH_URL } from '~/helpers/consts'
+import { ContentContext, HistoryContext } from '~/helpers/contexts'
+import { LINE_HOTKEYS } from '~/helpers/keyMap'
+import { findLineIndex } from '~/helpers/line'
+import { useCurrentLines } from '~/hooks'
+import controller from '~/services/controller'
+
 import ShabadInfo from './ShabadInfo'
 import ToolbarButton from './ToolbarButton'
