+++ apps/frontend/src/services/analytics.ts
@@ -4,10 +4,10 @@
  */
 
 /* eslint-disable class-methods-use-this */
-import { About } from '@presenter/contract/src/about'
+import { About } from '@presenter/contract'
 import * as Sentry from '@sentry/browser'
 
-import { BACKEND_URL, isDev, SENTRY_DSN, SENTRY_PROJECT } from '../helpers/consts'
+import { BACKEND_URL, isDev, SENTRY_DSN, SENTRY_PROJECT } from '~/helpers/consts'
 
 /**
  * Analytics class for tracking events and providing error reporting.
