+++ packages/contract/src/index.ts
@@ -1,3 +1,4 @@
+export * from './about'
 export * from './data'
 export * from './settings'
 export * from './websocket'
